<fix> force convert to double

+++ cougar/cext/median_heap_impl.h
@@ -250,9 +250,9 @@ static inline void method(pop, T)(median_heap(T) * heap) {
 
 static inline double method(query_median, T)(median_heap(T) * heap) {
     if (heap->min_size == heap->max_size) {
-        return (heap->min_heap[0]->value + heap->max_heap[0]->value) * 0.5;
+        return (double)(heap->min_heap[0]->value + heap->max_heap[0]->value) * 0.5;
     } else {
-        return heap->min_heap[0]->value;
+        return (double)(heap->min_heap[0]->value);
     }
 }
